requirements.txt
flask==2.0.1
werkzeug==2.0.3
google-generativeai==0.3.2 